requirements.txt
Django==1.7
djangorestframework==2.4.1
MySQL-python==1.2.5
Werkzeug==0.9.6
bpython==0.13.1
django-extensions==1.3.11
